+++ src/modules/Elsa.Workflows.Core/Extensions/ActivityExtensions.cs
@@ -64,7 +64,7 @@ public static class ActivityExtensions
     /// <param name="context">The activity execution context.</param>
     /// <param name="outputName">Name of the output.</param>
     /// <returns>The output value.</returns>
-    public static object? GetOutput(this IActivity activity, ActivityExecutionContext context, string? outputName = default)
+    public static object? GetOutput(this IActivity activity, ActivityExecutionContext context, string? outputName = null)
     {
         var workflowExecutionContext = context.WorkflowExecutionContext;
         var outputRegister = workflowExecutionContext.GetActivityOutputRegister();
@@ -84,11 +84,11 @@ public static class ActivityExtensions
     /// <param name="context">The expression execution context.</param>
     /// <param name="outputName">Name of the output.</param>
     /// <returns>The output value.</returns>
-    public static object? GetOutput(this IActivity activity, ExpressionExecutionContext context, string? outputName = default)
+    public static object? GetOutput(this IActivity activity, ExpressionExecutionContext context, string? outputName = null)
     {
         var activityExecutionContext = context.GetActivityExecutionContext();
 
-        if (activityExecutionContext == null)
+        if (activityExecutionContext == null!)
             return null;
         
         return activity.GetOutput(activityExecutionContext, outputName);
