+++ src/modules/Elsa.Workflows.Runtime/Middleware/Activities/EvaluateLogPersistenceModesMiddleware.cs
@@ -1,0 +1,15 @@
+using Elsa.Workflows.Pipelines.ActivityExecution;
+
+namespace Elsa.Workflows.Runtime.Middleware.Activities;
+
+public class EvaluateLogPersistenceModesMiddleware(ActivityMiddlewareDelegate next, IActivityPropertyLogPersistenceEvaluator persistenceEvaluator) : IActivityExecutionMiddleware
+{
+    internal static object LogPersistenceMapKey { get; } = new();
+    
+    public async ValueTask InvokeAsync(ActivityExecutionContext context)
+    {
+        await next(context);
+        var persistenceLogMap = persistenceEvaluator.EvaluateLogPersistenceModesAsync(context);
+        context.TransientProperties[LogPersistenceMapKey] = persistenceLogMap;
+    }
+}
