+++ src/modules/Elsa.Workflows.Runtime/Middleware/Workflows/PersistWorkflowExecutionLogMiddleware.cs
@@ -13,8 +13,5 @@ public class PersistWorkflowExecutionLogMiddleware(WorkflowMiddlewareDelegate ne
     {
         // Invoke next middleware.
         await Next(context);
-
-        // Not used anymore.
-        //await sink.PersistExecutionLogsAsync(context);
     }
 }
